manager_new/pyproject.toml
[tool.poetry]
name = "ad_db_manager"
version = "0.1.0"
description = "Data manager for maintaining vehicle data from different sources and create an up-to-date list of information (eg. OBD DTCs)"
authors = ["autodiag2 <autodiag@netcourrier.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
toml = "^0.10.2"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
pyyaml = "^6.0.3"
jinja2 = "^3.1.6"
python-multipart = "^0.0.22"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ad_db_manager = 'ad_db_manager.main:main'